Move `WC_Google_Gtag_JS` tests to a separate class
as we will add more and more to check scripts registration behavior.
